+++ otx/algorithms/visual_prompting/adapters/pytorch_lightning/datasets/__init__.py
@@ -1,0 +1,9 @@
+"""OTX Algorithms - Visual prompting Dataset."""
+
+# Copyright (C) 2023 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+
+from .dataset import OTXVisualPromptingDataModule
+from .pipelines import MultipleInputsCompose, Pad, ResizeLongestSide
+
+__all__ = ["OTXVisualPromptingDataModule", "ResizeLongestSide", "MultipleInputsCompose", "Pad"]