+++ otx/algorithms/visual_prompting/adapters/pytorch_lightning/models/decoders/__init__.py
@@ -1,0 +1,8 @@
+"""Encoders for visual prompting model."""
+
+# Copyright (C) 2023 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+
+from .sam_mask_decoder import SAMMaskDecoder
+
+__all__ = ["SAMMaskDecoder"]