+++ tests/unit/core/data/adapter/test_init.py
@@ -22,6 +22,8 @@ def test_get_dataset_adapter_incremental(task_name, train_type):
     root_path = os.getcwd()
     task_type = TASK_NAME_TO_TASK_TYPE[task_name]
     data_root = TASK_NAME_TO_DATA_ROOT[task_name]
+    if str(task_type).upper() == "VISUAL_PROMPTING":
+        data_root = data_root.get("coco")
 
     get_dataset_adapter(
         task_type=task_type,
@@ -105,7 +107,7 @@ def test_direct_annotation(task_name, train_type):
         train_ann_files="tests/assets/car_tree_bug/annotations/instances_train_5_imgs.json",
         val_data_roots=os.path.join(root_path, data_root["val"]),
     )
-    assert t_adapter.dataset[Subset.TRAINING].get_subset("train").get_annotated_items() == 5
+    assert t_adapter.dataset[Subset.TRAINING].get_subset("train_5_imgs").get_annotated_items() == 5
 
     v_adapter = get_dataset_adapter(
         task_type=task_type,
@@ -114,7 +116,7 @@ def test_direct_annotation(task_name, train_type):
         val_data_roots=os.path.join(root_path, data_root["val"]),
         val_ann_files="tests/assets/car_tree_bug/annotations/instances_val_1_imgs.json",
     )
-    assert v_adapter.dataset[Subset.VALIDATION].get_subset("val").get_annotated_items() == 1
+    assert v_adapter.dataset[Subset.VALIDATION].get_subset("val_1_imgs").get_annotated_items() == 1
 
     tv_adapter = get_dataset_adapter(
         task_type=task_type,
@@ -124,8 +126,8 @@ def test_direct_annotation(task_name, train_type):
         val_data_roots=os.path.join(root_path, data_root["val"]),
         val_ann_files="tests/assets/car_tree_bug/annotations/instances_val_1_imgs.json",
     )
-    assert tv_adapter.dataset[Subset.TRAINING].get_subset("train").get_annotated_items() == 5
-    assert tv_adapter.dataset[Subset.VALIDATION].get_subset("val").get_annotated_items() == 1
+    assert tv_adapter.dataset[Subset.TRAINING].get_subset("train_5_imgs").get_annotated_items() == 5
+    assert tv_adapter.dataset[Subset.VALIDATION].get_subset("val_1_imgs").get_annotated_items() == 1
 
 
 @e2e_pytest_unit